WIP implementing server test with audio file
